# frontend/openapi/return-platform.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Return Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/data-console/v1/scenarios": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_data_console_v1_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_Scenario__"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Create Scenario",
        "operationId": "create_scenario_data_console_v1_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Scenario_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/scenarios/{scenario_id}": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Get Scenario",
        "operationId": "get_scenario_data_console_v1_scenarios__scenario_id__get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Scenario_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/scenarios/{scenario_id}/generate": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Generate Scenario",
        "operationId": "generate_scenario_data_console_v1_scenarios__scenario_id__generate_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Scenario_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/scenarios/{scenario_id}/validate": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Validate Scenario",
        "operationId": "validate_scenario_data_console_v1_scenarios__scenario_id__validate_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Scenario_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/scenarios/{scenario_id}/approve": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Approve Scenario",
        "operationId": "approve_scenario_data_console_v1_scenarios__scenario_id__approve_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_Scenario_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/data-console/v1/scenarios/{scenario_id}/preview": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "summary": "Preview Scenario",
        "operationId": "preview_scenario_data_console_v1_scenarios__scenario_id__preview_get",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_Scenario_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scenario"
              },
              {
                "type": "null"
              }
            ]
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "APIResponse[Scenario]"
      },
...
      "APIResponse_list_Scenario__": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Scenario"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "page": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "meta"
        ],
        "title": "APIResponse[list[Scenario]]"
      },
...
      "CreateScenarioPayload": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "baseWorkspaceId": {
            "type": "string",
            "title": "Baseworkspaceid"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "baseWorkspaceId",
          "parameters"
        ],
        "title": "CreateScenarioPayload"
      },
...
      "FreshnessStatus": {
        "type": "string",
        "enum": [
          "LIVE",
          "CACHED",
          "STALE"
        ],
        "title": "FreshnessStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "page_size": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 50
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PageMeta"
      },
      "ResponseMeta": {
        "properties": {
          "schema_version": {
            "type": "string",
            "pattern": "^\\d+\\.\\d+$",
            "title": "Schema Version",
            "default": "1.0"
          },
          "request_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Request Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessStatus",
            "default": "LIVE"
          },
          "partial": {
            "type": "boolean",
            "title": "Partial",
            "default": false
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningMeta"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta"
      },
...
      "Scenario": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "baseWorkspaceId": {
            "type": "string",
            "title": "Baseworkspaceid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "baseWorkspaceId",
          "status",
          "parameters",
          "createdAt",
          "owner"
        ],
        "title": "Scenario"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WarningMeta": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Source"
          },
          "code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "code",
          "message"
        ],
        "title": "WarningMeta"
      },
...
    }
  }
}